+++ noodles-util/src/variant/async/io/reader/inner.rs
@@ -6,6 +6,8 @@ use noodles_bgzf as bgzf;
 use noodles_vcf as vcf;
 use tokio::io::{self, AsyncRead, BufReader};
 
+use crate::variant::Record;
+
 pub(super) enum Inner<R>
 where
     R: AsyncRead,
@@ -26,6 +28,55 @@ where
             Self::BcfRaw(reader) => reader.read_header().await,
             Self::Vcf(reader) => reader.read_header().await,
             Self::VcfGz(reader) => reader.read_header().await,
+        }
+    }
+
+    pub(super) async fn read_record(&mut self, record: &mut Record) -> io::Result<usize> {
+        match self {
+            Inner::Bcf(reader) => {
+                if !matches!(record, Record::Bcf(_)) {
+                    *record = Record::Bcf(bcf::Record::default());
+                }
+
+                if let Record::Bcf(r) = record {
+                    reader.read_record(r).await
+                } else {
+                    unreachable!();
+                }
+            }
+            Inner::BcfRaw(reader) => {
+                if !matches!(record, Record::Bcf(_)) {
+                    *record = Record::Bcf(bcf::Record::default());
+                }
+
+                if let Record::Bcf(r) = record {
+                    reader.read_record(r).await
+                } else {
+                    unreachable!();
+                }
+            }
+            Inner::Vcf(reader) => {
+                if !matches!(record, Record::Vcf(_)) {
+                    *record = Record::Vcf(vcf::Record::default());
+                }
+
+                if let Record::Vcf(r) = record {
+                    reader.read_record(r).await
+                } else {
+                    unreachable!();
+                }
+            }
+            Inner::VcfGz(reader) => {
+                if !matches!(record, Record::Vcf(_)) {
+                    *record = Record::Vcf(vcf::Record::default());
+                }
+
+                if let Record::Vcf(r) = record {
+                    reader.read_record(r).await
+                } else {
+                    unreachable!();
+                }
+            }
         }
     }