+++ noodles-util/CHANGELOG.md
@@ -13,6 +13,13 @@
 
 [#348]: https://github.com/zaeleus/noodles/pull/348
 
+### Added
+
+  * util/variant/async/io/reader: Add variant record reader
+    (`Reader::read_record`) ([#349]).
+
+[#349]: https://github.com/zaeleus/noodles/pull/349
+
 ## 0.70.0 - 2025-08-25
 
 ### Added